backend/requirements.txt
anthropic
langchain-text-splitters
chromadb
fastapi
uvicorn[standard]
python-dotenv
PyPDF2
python-multipart
requests
beautifulsoup4
python-jose[cryptography]
passlib[bcrypt]
